+++ RELEASE_NOTES.md
@@ -25,6 +25,8 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
     - In case your ASCOM Alpaca specific device has a static IP or doesn't offer Alpaca Discovery a new static entry is available for each device type to pick from where you can specify the address to connect to instead of having to rely on discovery
 - **Altair, Mallincam, Ogma, Omegon, Risingcam, SvBony and ToupTek Filterwheel Native Driver**
   - The ToupTek based filter wheels are now available as a native driver.
+- **Oasis Focuser Native Driver**
+  - The Oasis focuser is now available as a native driver.
 
 # Version 3.2